Create Exercise 3-4
Summary
During each summer, John and Jessica grow vegetables in their backyard and buy seeds and fertilizer from a local nursery. The nursery carries different types of vegetable fertilizers in various bag sizes.

When buying a particular fertilizer, they want to know the price of the fertilizer per pound and the cost of fertilizing per square foot.

Instructions
The following program prompts the user to enter the size of the fertilizer bag, in pounds, the cost of the bag, and the area, in square feet, that can be covered by the bag.

The program should output the desired result. However, the program contains logic errors. Find and correct the logic errors so that the program works properly.